+++ lib/services/injections.ts
@@ -1,5 +1,5 @@
-import { makeFariFirebase } from "./firebase/firebase";
 import { InternationalizationService } from "./internationalization/InternationalizationService";
+import { makeLiveBlocksClient as makeLiveBlocksClient } from "./liveblocks/makeLiveBlocksClient";
 import { makeLogger } from "./logger/makeLogger";
 import { makeSentryService } from "./sentry/SentryService";
 
@@ -7,8 +7,14 @@ export function getDefaultInjections() {
   const sentryService = makeSentryService();
   const logger = makeLogger(sentryService);
   const internationalizationService = InternationalizationService(logger);
-  const fariFirebase = makeFariFirebase();
-  return { internationalizationService, logger, sentryService, fariFirebase };
+  const liveBlocksClient = makeLiveBlocksClient();
+
+  return {
+    internationalizationService,
+    logger,
+    sentryService,
+    liveBlocksClient,
+  };
 }
 
 export type IInjections = ReturnType<typeof getDefaultInjections>;